+++ contracts/registry/src/lib.rs
@@ -1,52 +1,19 @@
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
-use near_sdk::collections::{LookupMap, UnorderedMap, UnorderedSet};
-use near_sdk::env::STORAGE_PRICE_PER_BYTE;
-use near_sdk::json_types::{Base64VecU8, U128, U64};
+use near_sdk::collections::{LookupMap, UnorderedSet};
 use near_sdk::serde::{Deserialize, Serialize};
-use near_sdk::{
-    env, ext_contract, log, near_bindgen, require, serde_json, AccountId, Balance, BorshStorageKey,
-    Gas, Promise, PromiseError,
-};
+use near_sdk::{env, near_bindgen, AccountId, BorshStorageKey};
 
+pub mod admins;
 pub mod internal;
+pub mod projects;
 pub mod utils;
+pub use crate::admins::*;
 pub use crate::internal::*;
+pub use crate::projects::*;
 pub use crate::utils::*;
 
 type ProjectId = AccountId;
 type TimestampMs = u64;
-
-#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, PartialEq, Debug)]
-#[serde(crate = "near_sdk::serde")]
-pub enum ProjectStatus {
-    Submitted,
-    InReview,
-    Approved,
-    Rejected,
-}
-
-#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, PartialEq, Debug)]
-#[serde(crate = "near_sdk::serde")]
-pub struct ProjectInternal {
-    pub id: ProjectId,
-    pub name: String,
-    pub status: ProjectStatus,
-    pub submitted_ms: TimestampMs,
-    pub updated_ms: TimestampMs,
-    pub review_notes: Option<String>,
-}
-
-#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, PartialEq, Debug)]
-#[serde(crate = "near_sdk::serde")]
-pub struct ProjectExternal {
-    pub id: ProjectId,
-    pub name: String,
-    pub team_members: Vec<AccountId>,
-    pub status: ProjectStatus,
-    pub submitted_ms: TimestampMs,
-    pub updated_ms: TimestampMs,
-    pub review_notes: Option<String>,
-}
 
 /// Registry Contract
 #[near_bindgen]
@@ -55,8 +22,22 @@ pub struct Contract {
     owner: AccountId,
     admins: UnorderedSet<AccountId>,
     project_ids: UnorderedSet<ProjectId>,
-    projects_by_id: LookupMap<ProjectId, ProjectInternal>,
+    projects_by_id: LookupMap<ProjectId, VersionedProjectInternal>,
     project_team_members_by_project_id: LookupMap<ProjectId, UnorderedSet<AccountId>>,
+}
+
+#[derive(BorshSerialize, BorshDeserialize)]
+pub enum VersionedContract {
+    Current(Contract),
+}
+
+/// Convert VersionedContract to Contract
+impl From<VersionedContract> for Contract {
+    fn from(contract: VersionedContract) -> Self {
+        match contract {
+            VersionedContract::Current(current) => current,
+        }
+    }
 }
 
 #[derive(BorshSerialize, BorshStorageKey)]
@@ -83,112 +64,8 @@ impl Contract {
             ),
         }
     }
-
-    #[payable]
-    pub fn owner_add_admins(&mut self, admins: Vec<AccountId>) {
-        self.assert_owner();
-        for admin in admins {
-            self.admins.insert(&admin);
-        }
-    }
-
-    pub fn get_admins(&self) -> Vec<AccountId> {
-        self.admins.to_vec()
-    }
-
-    #[payable]
-    pub fn owner_remove_admins(&mut self, admins: Vec<AccountId>) {
-        self.assert_owner();
-        for admin in admins {
-            self.admins.remove(&admin);
-        }
-    }
-
-    #[payable]
-    pub fn register(
-        &mut self,
-        name: String,
-        team_members: Vec<AccountId>,
-        _project_id: Option<AccountId>,
-    ) -> ProjectExternal {
-        // TODO: require enough funds to cover storage deposit
-        // _project_id can only be specified by admin; otherwise, it is the caller
-        let project_id = if let Some(_project_id) = _project_id {
-            self.assert_admin();
-            _project_id
-        } else {
-            env::predecessor_account_id()
-        };
-        self.assert_project_does_not_exist(&project_id);
-        let project_internal = ProjectInternal {
-            id: project_id.clone(),
-            name,
-            status: ProjectStatus::Approved, // approved by default - TODO: double-check that this is desired functionality
-            submitted_ms: env::block_timestamp_ms(),
-            updated_ms: env::block_timestamp_ms(),
-            review_notes: None,
-        };
-        self.project_ids.insert(&project_id);
-        self.projects_by_id.insert(&project_id, &project_internal);
-        let mut team_members_set =
-            UnorderedSet::new(StorageKey::ProjectTeamMembersByProjectIdInner {
-                project_id: project_id.clone(),
-            });
-        for team_member in team_members {
-            team_members_set.insert(&team_member);
-        }
-        self.project_team_members_by_project_id
-            .insert(&project_id, &team_members_set);
-        self.format_project(project_internal)
-    }
-
-    pub fn get_projects(&self) -> Vec<ProjectExternal> {
-        self.project_ids
-            .iter()
-            .map(|project_id| {
-                self.format_project(self.projects_by_id.get(&project_id).expect("No project"))
-            })
-            .collect()
-    }
-
-    pub fn get_project_by_id(&self, project_id: ProjectId) -> ProjectExternal {
-        self.format_project(self.projects_by_id.get(&project_id).expect("No project"))
-    }
-
-    pub(crate) fn format_project(&self, project_internal: ProjectInternal) -> ProjectExternal {
-        ProjectExternal {
-            id: project_internal.id.clone(),
-            name: project_internal.name,
-            team_members: self
-                .project_team_members_by_project_id
-                .get(&project_internal.id)
-                .unwrap()
-                .to_vec(),
-            status: project_internal.status,
-            submitted_ms: project_internal.submitted_ms,
-            updated_ms: project_internal.updated_ms,
-            review_notes: project_internal.review_notes,
-        }
-    }
-
-    #[payable]
-    pub fn admin_set_project_status(
-        &mut self,
-        project_id: ProjectId,
-        status: ProjectStatus,
-        review_notes: Option<String>,
-    ) {
-        self.assert_admin();
-        self.assert_project_exists(&project_id);
-        let mut project = self.projects_by_id.get(&project_id).unwrap();
-        project.status = status;
-        project.review_notes = review_notes;
-        project.updated_ms = env::block_timestamp_ms();
-        self.projects_by_id.insert(&project_id, &project);
-    }
 }
 
-// TODO: not sure why this is necessary
 impl Default for Contract {
     fn default() -> Self {
         Self {
