+++ contracts/test/registry/config.ts
@@ -1,3 +1,3 @@
-export const contractId = "dev-1695147323072-34494206941153";
+export const contractId = "dev-1696953571615-36687499584323";
 export const networkId = "testnet";
 export const nodeUrl = `https://rpc.${networkId}.near.org`;
